requirements.txt
python-dotenv>=1.0.0
yt-dlp>=2024.5.26
pydub>=0.25.1
discord.py>=2.3.2
openai>=1.21.0
httpx>=0.27.0
